requirement.txt
opentelemetry-api==1.15.0
opentelemetry-sdk==1.15.0